deduplicate text formatting files progress table expando
